requirements.txt
# Core ML
torch
transformers==4.37.2
sentence-transformers==2.2.2
huggingface-hub==0.19.4

langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-google-genai==1.0.8
langchain-pinecone==0.1.3
langchain-text-splitters==0.0.1

# Backend
fastapi
uvicorn

# Vector DB
pinecone
orjson

# Utils
pypdf
python-dotenv
openai
